refactor: remove old files
