+++ .vscode/settings.json
@@ -12,5 +12,6 @@
     "string_view": "c",
     "sstream": "c",
     "random": "c"
-  }
+  },
+  "C_Cpp.errorSquiggles": "enabled"
 }
