+++ src/logger.h
@@ -1,10 +1,5 @@
 #pragma once
-
 #include <stdio.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
 
 #ifndef PRETTY_LOG
 #define PRETTY_LOG 1
@@ -24,7 +19,3 @@ extern "C" {
     #define log_warn(fmt, ...)  printf("[WARN] " fmt "\n", ##__VA_ARGS__)
     #define log_err(fmt, ...) fprintf(stderr, "[ERROR] " fmt "\n", ##__VA_ARGS__)
 #endif
-
-#ifdef __cplusplus
-} // extern "C"
-#endif // __cplusplus
