--- src/luxio.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "common.h"
-#include "event.h"
-#include "net.h"
-#include "timer.h"
