+++ src/pandio.h
@@ -1,4 +1,3 @@
-#include <cstdint>
 #include <stdint.h>
 #include "common.h"
 #include "heap.h"
