+++ src/timers.c
@@ -1,4 +1,4 @@
-#include "timers2.h"
+#include "timers.h"
 #include "heap.h"
 #include <limits.h>
 
@@ -6,7 +6,7 @@
 #define pnd_timer_unwrap(ptr) \
   (ptr == NULL) ? NULL : container_of(ptr, pnd_timer_t, hnode)
 
-/* min-heap - answers question: should swap? */
+/* answers question: should swap? */
 int pnd_timers_comparator(struct heap_node *a, 
                           struct heap_node *b) {
   pnd_timer_t *child = pnd_timer_unwrap(a);
@@ -15,13 +15,11 @@ int pnd_timers_comparator(struct heap_node *a,
   return child->timeout < parent->timeout;
 }
 
-void pnd_timers_heap_init(pnd_io_t *ctx)
-{
+void pnd_timers_heap_init(pnd_io_t *ctx) {
   heap_init(&ctx->timers, pnd_timers_comparator);  
 }
 
-void pnd_timer_init(pnd_io_t *ctx, pnd_timer_t *timer)
-{
+void pnd_timer_init(pnd_io_t *ctx, pnd_timer_t *timer) {
   timer->ctx = ctx;
   timer->timeout = 0;
   timer->interval = 0;
@@ -51,8 +49,7 @@ void pnd_timer_repeat(pnd_timer_t *timer,
   pnd_timer_start(timer, cb, interval);
 }
 
-void pnd_timer_stop(pnd_timer_t *timer)
-{
+void pnd_timer_stop(pnd_timer_t *timer) {
   if (!timer->active) {
     return;
   }
@@ -73,7 +70,10 @@ void pnd_timers_run(pnd_io_t *ctx) {
       break;
 
     pnd_timer_stop(min);
-    min->on_timeout(min);
+
+    // very unlikely scenario to be NULL, but still...
+    if (min->on_timeout)
+      min->on_timeout(min);
 
     if (min->interval != 0)
       pnd_timer_start(min, min->on_timeout, min->interval);
